Cleaning up templates

+++ bin/dev-init/util.ts
@@ -10,11 +10,17 @@ export class Util {
   static F_ROOT = F_ROOT;
 
   static copyTemplateFiles(src: string, base: string) {
+    const module = path.basename(base);
     for (const f of fs.readdirSync(src)) {
-      const srcFile = `${src}/${f}`;
-      const destFile = `${base}/${f}`;
-      if (fs.statSync(srcFile).isFile() && fs.existsSync(destFile)) {
-        fs.copyFileSync(srcFile, destFile);
+      const resolved = path.resolve(src, f);
+      if (fs.statSync(resolved).isFile()) {
+        const resolvedModule = path.resolve(src, module, f);
+        const srcFile = fs.existsSync(resolvedModule) ? resolvedModule : resolved;
+        const destFile = path.resolve(base, f);
+
+        if (fs.statSync(srcFile).isFile() && fs.existsSync(destFile)) {
+          fs.copyFileSync(srcFile, destFile);
+        }
       }
     }
   }
